# api/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobResultUrls": {
        "additionalProperties": false,
        "description": "Artifact URLs produced by a successful transcription job.",
        "properties": {
          "midi": {
            "title": "Midi",
            "type": "string"
          },
          "musicxml": {
            "title": "Musicxml",
            "type": "string"
          },
          "pdf": {
            "title": "Pdf",
            "type": "string"
          }
        },
        "required": [
          "musicxml",
          "midi",
          "pdf"
        ],
        "title": "JobResultUrls",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "pending",
          "running",
          "done",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobStatusResponse": {
        "additionalProperties": false,
        "description": "Response body for GET /jobs/{job_id}.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "progress": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "result_urls": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobResultUrls"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "job_id",
          "status",
          "progress"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
      "TranscribeRequest": {
        "additionalProperties": false,
        "description": "Request body for POST /transcribe.\n\nExactly one of `audio_url` or `audio_file_b64` must be provided.",
        "properties": {
          "audio_file_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Base64-encoded audio payload (mp3/wav).",
            "title": "Audio File B64"
          },
          "audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "HTTP(S) URL the backend will fetch the audio from.",
            "title": "Audio Url"
          }
        },
        "title": "TranscribeRequest",
        "type": "object"
      },
      "TranscribeResponse": {
        "additionalProperties": false,
        "description": "Response body for POST /transcribe.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "pending"
          }
        },
        "required": [
          "job_id",
          "created_at"
        ],
        "title": "TranscribeResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Audio-to-lead-sheet transcription service. Endpoints are stubbed until ML inference (Agent C) and the job queue (PR-14) land.",
    "title": "music2sheet API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Healthz Healthz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz"
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/transcribe": {
      "post": {
        "operationId": "create_transcription_job_transcribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Transcription Job"
      }
    }
  }
}